Split OpDetailForm Date field into 2 fields
A fix for Date range picker calendar picker overflow the screen. The reason is that the current Antd Design is not compatible with mobile device. Adding antd design will add a lot of storage so it is not good. The second option would be split the date field into start date and end date. The schema is unchanged