VP-480: Change tags to be plain strings instead of a separate Mongo doc
Make Activities, People and Opportunities use strings for tags. This
make editing/managing the documents simpler. A separate document
with all the candidate tags will be maintained and updated each time
one of the taggable documents is updated.

The full list of potential tags will be loaded into the client for
autocompletion.