VP-527 Handle null person field in InterestTable
This should almost never happen but in the event it does we shouldn't
show interest records where the person field is null. So we are now
filtering those out when giving the data to the Table component.

I've also updated the interests prop on InterestTable to be required -
it looks like it should always be provided (even if it happens to be an
empty array). I've updated some unit tests to match this assumption.