VP-908 Fix message flash when creating new org
When submitting the org detail form the user would briefly see a message
saying "Sorry this organisation is not available" before ending up on
the new organisation page. This appears to be because of a weird timing
issue with state/props on the page where you could still have the
"isNew" prop set to true while also having an organisation loading in
this.props.organisations.data.