VP-1095 Fix getInitialProps on InviteSchool page
Because getInitialProps can be called on the client or server side
sometimes the API request can be made from the server side and will be
sent with no cookies. These requests were failing because of the newly
enforced permissions (and the fact that authentication relies on the
cookies being in the request).

We now detect the req object and pass the cookies along in the event the
request is happening on the server side.