VP-82
All Contributors work through this initial ticket to demonstrate that they can run the code and have basic understanding of the commit processes.

Assign this ticket to yourself and move it into ‘In Progress’

git clone the code base for vly2 - read the README, 

set up a developer environment either locally or using the docker containers.

demonstrate mongodb is working (visit http://localhost:27017 or http://localhost:3122/ for example)

run the tests

run the application

sign in

create a new opportunity and exercise the system

Now ask to be added to the developers team on github so that you can commit to the codebase.

create a git branch for this ticket

Read the CODE_OF_CONDUCT only continue if you accept these terms. 

Read and add your name and email to the contributors file

commit the change

issue a pull request. 

When PR is accepted and merged, delete the branch and move to your next ticket.